New client test module; Skip geth client signature integration test on CI until a better solution is found.
